ATR-762: renamed namespace for PX1009 and PX1011 analyzer
