+++ src/Acuminator/Acuminator.Analyzers/StaticAnalysis/Dac/DacAnalyzersAggregator.cs
@@ -12,7 +12,7 @@ using Acuminator.Analyzers.StaticAnalysis.DacPropertyAttributes;
 using Acuminator.Analyzers.StaticAnalysis.DacReferentialIntegrity;
 using Acuminator.Analyzers.StaticAnalysis.DacUiAttributes;
 using Acuminator.Analyzers.StaticAnalysis.ForbiddenFieldsInDac;
-using Acuminator.Analyzers.StaticAnalysis.InheritanceFromPXCacheExtension;
+using Acuminator.Analyzers.StaticAnalysis.DeclarationAnalysisDac;
 using Acuminator.Analyzers.StaticAnalysis.LegacyBqlField;
 using Acuminator.Analyzers.StaticAnalysis.MethodsUsageInDac;
 using Acuminator.Analyzers.StaticAnalysis.MissingBqlFieldRedeclarationInDerived;
